+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This API sends notifications to users for things like forgotten passwords, initi
 ## unreleased
 ### changed
 - YLP-682 Add team information in received invitations
+### Fixed 
+- YLP-709 Bug: get the coorect ID to dismiss an invite
+
 ## 1.6.1 - 2021-05-04
 ### Changed
 - YLP-710 Modify route /accept/team/invite to acknowledge notifications that does not requires a specific action (i.e. change member role)